edited layout

+++ features/noise/src/main/java/com/example/noise/ui_noise/model/FrequencyState.kt
@@ -1,5 +1,5 @@
 package com.example.noise.ui_noise.model
 
 data class FrequencyState(
-    val frequencies: FloatArray
+    val frequencies: DoubleArray
 )
